Putricide: remove Malleable Goo target scanner
On Warmane, this method does not work. This is due to the boss never swapping targets to throw the malleable, like it happens on retail, causing the addon to assume that the tank was the malleable target every time, which is false.
After transcriptor log reviews, there is no indication that a way to check for this is even possible with Warmane's current boss script.