Core: fix UNIT event filtering
Mod registered unit events that had args (like boss1 boss2 boss3 etc) were being sequentially shortened during handling based on the previous mod registered unit event if it also had matching args.
To circumvent this, a current encounter variable (currentModId) was added to StartCombat, and nil'd on EndCombat. This variable is now being check on the event handler to ensure only the specific mod events that were registered on the boss lua are handled.

Thanks to Kader and Barsoom for their time and support on this fix.

Co-Authored-By: Kader B <bkader@mail.com>