Core: fix EndCombat message showing World Boss if called outside of instance
On a wipe, EndCombat was often being called outside of the instance due to spirit being released before EndCombat finished the schedule, where SecondaryLoadCheck overwrites cached difficulty variables with what's current on the zone (since it reruns GetCurrentInstanceDifficulty).
To overcome this, new variables are now created on StartCombat to register the encounter related difficulties and use those instead for EndCombat purposes, wiping them at the end of the EndCombat logic.