Core: fix UnregisterInCombatEvents for unit events
mod.inCombatOnlyEvents table contains untreated "events" from the module (event + filter). On the other hand, registeredEvents is keyed by the proper event that is registered. For this reason, UnregisterInCombatEvents was failing to unregister unit events because checkEntry was returning false (example: UNIT_HEALTH ~= UNIT_HEALTH boss1).

This new logic fixes the problem by splitting the untreated the event, which enables the right check to validate the event in the table to be tremoved.