PVP: fix Focused Assault timer
Brutal assault would be too costly for the 0.01% chance of it happening in WG. If the server has the chat strings scripted, then it will work, otherwise, there won't be anything in the code handling it.

This code is not very robust either: once the buff applies, everytime the flag is dropped and gets picked up again, the timer will start back to 10min.