BPC: add Empowered Shock Vortex timer (Fix #69)
There is an inherent random factor associated with these timers. The only timer that is reproducible is Empowered Shock Vortex to Empowered Shock Vortex, or Empowered Shock Vortex to the Shock Vortex right after it (30 seconds). Everything else fluctuates, based on logs.

To mitigate the random factor, I have passing the Keep argument as true, so the bar stays on screen until cancelled of restarted.

Additionally, I am checking bar remaining timers and started status, which might not be needed at all with the Keep now enabled.