PVP: remake faction checker for Cross-Faction BGs
Apparently, not every server employs the learning of the default faction Language that you're playing for.

New logic queries server on pvp instance init, with populates the BGScore for the API to return values. I am not registering nor Ace repeating this server query, based on the assumption that I will only need to query once for the faction checker to have at least the player.
Also something to keep in mind: ElvUI users will always have up to date info, since its API has a RepeatingSchedule of this query.