SSC-Tidewalker: fix Watery Grave timer
Warmane bugged this mechanic (as of 2022/08/11) and it's not triggering the ability at random intervals. The emote still fires every 30 seconds, so use that for timer (even though the ability may never happen).
Added review flag to this timer as well since there was a 2s blip on one timer...