SSC-Vashj: new timers + magic barrier experimental + Tainted core bugfix
Added CD timers for P1/P3 spells, even though they have a considerable variance (10 seconds). "Keep" arg passed on these timers to visualize variance.

Magic Barrier does not have a CLEU event that we can listen to. From test server I did notice UNIT_SPELLCAST_FAILED_QUIET showing on event trace whenever a player used the Tainted Core on the generators - to be confirmed if it also happens on Warmane.

Lastly, CHAT_MSG_LOOT string differs whether it's the player or a group member that loots the item, which was passing "You" on EN clients with the coded string.match. I applied an attempt to catch these localized "You" and assume it's the player, or failsafe to player name regardless for locales that don't follow these rules or don't directly translate a "You" (ES, RU, ...).