LK: intermission staging + fix shambling enrage + adjust timers
- Now adds intermissions to the phasing model for this fight, which helps to get more accurate data.

- Shambling enrage apparently has a mechanic on Warmane where it can be skipped if stunned before cast. There's only hearsay evidence supporting this claim (local testing only showed that the enrage cast gets pushed until stun expires). Nevertheless, I have adjusted the code to do scheduling for this CD timer just in case, which shouldn't affect how the code works if this claim turns out to be untrue.
There was also a bug where the Announce for this mechanic was firing both on Cast Start as well as Aura Applied, so I opted to run this on Cast Start for more preemptive reactions.
Also added a differentiator count for this timer - spawn order - in order to be able to distinguish multiple Enrage CD bars when more than one shambling is alive.

- Several timers were adjusted based on log data.