Core: prevent possible client crash with Announces
There are some fonts that cause a full client crash if the Fontstring reaches a certain FontStringHeight and is escaped with textures. Although DBM Announcements were triggering this due to its grow animation, I could repro this with no addons and thus evidence indicates this is either a WotLK bug or a font problem.

⛔ PROBLEMATIC FONTS! Arm Wrestler, Ancient Geek, Capitalis Type Oasis, Germanica Embossed, Germanica Fluted, Germanica Shadowed, King Arthur Legend