ICC-Festergut: Adjust Gas Spore CD timers
This timer is a random mess - more variance on 2022 logs, and a bit more stable on 2023 logs, across 27 logs.
Tried to make it somewhat true to what happens the most, but the variance is impossible to account for as a NextTimer.
Lastly, old DBM assumed every third cast it would be an extra 10 seconds, but this was confirmed to be false on the logs, and therefore dismissed (only one log out of the 27 reached a 7th cast, and that was 50s apart from the 6th cast).