Ulduar-Thorim: rework Hard Mode
SPELL_AURA_REMOVED for Touch of Dominion was not working as expected, since Sif will always lose the debuff and fire this event.
Implemented Hard Mode detection based of the Yell, and added a new Special Announce to alert when Hard mode activates. GUI was also grouped into the respective achievement (25 player is a caveat, but it's only visual).