Realm: conditional mod load + Vanilla Onyxia
In the case Warmane starts diverging Onyxia realm boss scripts too much from its WotLK realm counterparts, then it might be sensible to have specific mods for Onyxia realm and a way to conditionally load them based on Realm name, while not having to specify Realm names for the "standard" mods.

One key note that was heavily deliberated was whether or not it would make to split the repositories, and the reason I decided against that idea was that Warmane still uses the same 3.3.5a client, therefore the addon folder will also be the same, hence it makes sense to ensure minimal impact for users who realm hop.