Core: mod culling will not be implemented
On Retail there was an extensive job of combining all raids into one folder, same with dungeons, even though there are significant downfalls like increased load times, loss of statistics and optimization decrements due to how DBM is coded to load everything under the folder.

As such, for this repository, I will keep everything as is since in my opinion it was a step in the wrong direction, and reworking the mod loading system to be selective on the lowest level of the hierarchy to accommodate the culling would be too much work.