BossBanner: try to catch only the gob name + reset encounterID on a timer
Gobs in general have yellow text in the first line of the tooltip. Since boss loot and their treasure chests are the main focus of this feature, attempt to cache only one line, yellow text, tooltip gobs.

Also, encounterID is being reset on a timer. Can be changed in the future.