BossBanner: show Loot animation if pending loot
Retail did not account for receiving Loot on the end of KILL animation, or that exceeds BB_MAX_LOOT. It would always require a new ENCOUNTER_LOOT_RECEIVED to iterate through pendingLoot.
This workaround now checks on the end of BossBanner animation if pendingLoot is populated in order to do another round of Loot animation with remaining loot.