ICC-LK: fix non-existing GUID on Soul Shriek timer (#264)
When LK casts Raging Spirit, it does not come with GUID, hence the timer was being fired without an identifier. This was causing the timer to be isolated from guid matching for Start and Cancel methods.
Although it is not conventional, I am forcing a reset of the timer with proper GUID once found by the player or group member targeting (will not sync this! Seems excessive for an add that is guaranteed to have at least someone targeting it).